requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
Flask-SocketIO==5.3.6
requests==2.32.3
gunicorn==22.0.0
bidict==0.23.1
blinker==1.8.2
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==43.0.0
dnspython==2.6.1
eventlet==0.36.1
gevent==24.2.1
greenlet==3.0.3
h11==0.14.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
numpy==2.0.1
packaging==24.1
pillow==10.4.0
pycparser==2.22
python-engineio==4.9.1
python-socketio==5.11.3
setuptools==69.1.1
simple-websocket==1.0.0
tqdm==4.66.4
urllib3==2.2.2
Werkzeug==3.0.3
wheel==0.43.0
wsproto==1.2.0
zope.event==5.0
zope.interface==6.4.post2
azure-storage-blob
azure-identity
prometheus_flask_exporter